Setup testing enviorement with jest and supertest
