+++ server/config/config.js
@@ -6,10 +6,7 @@ module.exports = {
     dialect: "postgres",
   },
   test: {
-    username: "root",
-    password: null,
-    database: "database_test",
-    host: "127.0.0.1",
+    use_env_variable: "TEST_DATABASE_URL",
     dialect: "postgres",
   },
   production: {
