+++ server/package.json
@@ -20,8 +20,10 @@
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "extract-date": "^2.8.2",
+    "jest": "^26.6.3",
     "pg": "^8.5.1",
     "sequelize": "^6.3.5",
-    "sequelize-cli": "^6.2.0"
+    "sequelize-cli": "^6.2.0",
+    "supertest": "6.0.0"
   }
 }
