+++ server/test/e2e.test.js
@@ -1,0 +1,7 @@
+const request = require("supertest");
+const app = require("../app");
+const db = require("../models");
+
+const server = request(app);
+
+test("should ", () => {});
